Move tests and add delay calls for test
FindAllScenariosData.test.js into __tests__ folder for convention purpose
GetScenariosTreeData.test.js into __tests__ folder for convention purpose
Move FindAllScenarios.json and GetScenariosTree.json into respective __tests__
