+++ src/state/sagas/scenario/GetScenariosTree/GetScenariosTreeData.js
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import axios from 'axios';
-import { put, takeEvery } from 'redux-saga/effects';
+import { delay, put, takeEvery } from 'redux-saga/effects';
 import { SCENARIO_ENDPOINT, SCENARIO_ACTIONS_KEY } from '../../../commons/ScenarioConstants';
 
 // generators function
@@ -10,6 +10,7 @@ export function * fetchScenarioTreeData () {
   try {
     // yield keyword is here to milestone and save the action
     const { data } = yield axios.get(SCENARIO_ENDPOINT.GET_SCENARIO_TREE);
+    yield delay(3000);
     // Here is an effect named put that indicate to the middleware that it can dispatch a SET_SCENARIO_TREE action with data as payload
     yield put({ type: SCENARIO_ACTIONS_KEY.SET_SCENARIO_TREE, tree: data });
   } catch (error) {
